chore(cli): Make baremetal safe to import in cli
